grandpa: extract out structs used for json serialization
