+++ client/finality-grandpa/rpc/src/lib.rs
@@ -17,13 +17,15 @@
 //! RPC API for GRANDPA.
 #![warn(missing_docs)]
 
-use finality_grandpa::BlockNumberOps;
-use sc_finality_grandpa::{AuthorityId, SharedAuthoritySet, SharedVoterState, report};
-
 use futures::{FutureExt, TryFutureExt};
 use jsonrpc_derive::rpc;
-use serde::{Deserialize, Serialize};
-use std::{collections::HashSet, fmt::Debug};
+use std::fmt::Debug;
+use finality_grandpa::BlockNumberOps;
+
+mod serialized;
+
+use sc_finality_grandpa::{SharedAuthoritySet, SharedVoterState};
+use serialized::ReportedRoundStates;
 
 /// Returned when Grandpa RPC endpoint is not ready.
 pub const NOT_READY_ERROR_CODE: i64 = 1;
@@ -32,7 +34,9 @@ type FutureResult<T> =
 	Box<dyn jsonrpc_core::futures::Future<Item = T, Error = jsonrpc_core::Error> + Send>;
 
 #[derive(derive_more::Display, derive_more::From)]
-enum Error {
+/// Top-level error type for the RPC handler
+pub enum Error {
+	/// The GRANDPA RPC endpoint is not ready.
 	#[display(fmt = "GRANDPA RPC endpoint not ready")]
 	EndpointNotReady,
 }
@@ -72,107 +76,6 @@ impl<Hash, Block> GrandpaRpcHandler<Hash, Block> {
 			shared_voter_state,
 			shared_authority_set,
 		}
-	}
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct Prevotes {
-	current_weight: u64,
-	missing: HashSet<AuthorityId>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct Precommits {
-	current_weight: u64,
-	missing: HashSet<AuthorityId>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RoundState {
-	round: u64,
-	total_weight: u64,
-	threshold_weight: u64,
-	prevotes: Prevotes,
-	precommits: Precommits,
-}
-
-impl RoundState {
-	fn from(
-		round: u64,
-		round_state: &report::RoundState<AuthorityId>,
-		voters: &HashSet<AuthorityId>,
-	) -> Self {
-		let prevotes = &round_state.prevote_ids;
-		let missing_prevotes = voters.difference(&prevotes).cloned().collect();
-
-		let precommits = &round_state.precommit_ids;
-		let missing_precommits = voters.difference(&precommits).cloned().collect();
-
-		Self {
-			round,
-			total_weight: round_state.total_weight,
-			threshold_weight: round_state.threshold_weight,
-			prevotes: Prevotes {
-				current_weight: round_state.prevote_current_weight,
-				missing: missing_prevotes,
-			},
-			precommits: Precommits {
-				current_weight: round_state.precommit_current_weight,
-				missing: missing_precommits,
-			},
-		}
-	}
-}
-
-/// The state of the current best round, as well as the background rounds in a
-/// form suitable for serialization.
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReportedRoundStates {
-	set_id: u64,
-	best: RoundState,
-	background: Vec<RoundState>,
-}
-
-impl ReportedRoundStates {
-	fn from<Hash, Block>(
-		voter_state: &SharedVoterState,
-		authority_set: &SharedAuthoritySet<Hash, Block>,
-	) -> Result<Self, Error>
-	where
-		Hash: Debug + Clone + Eq + Send + Sync + 'static,
-		Block: BlockNumberOps + Send + Sync + 'static,
-	{
-		let voter_state = voter_state
-			.voter_state()
-			.ok_or(Error::EndpointNotReady)?;
-
-		let current_voters: HashSet<AuthorityId> = authority_set
-			.current_authorities()
-			.voters()
-			.iter()
-			.map(|p| p.0.clone())
-			.collect();
-
-		let best = {
-			let (round, round_state) = voter_state.best_round;
-			RoundState::from(round, &round_state, &current_voters)
-		};
-
-		let background = voter_state
-			.background_rounds
-			.iter()
-			.map(|(round, round_state)| RoundState::from(*round, round_state, &current_voters))
-			.collect();
-
-		Ok(Self {
-			set_id: authority_set.set_id(),
-			best,
-			background,
-		})
 	}
 }
